requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
twilio==9.4.5
openai==1.68.2
